// apps/web/src/app/api/models/maintenance-record-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface MaintenanceRecordDto {
  costUsd?: number;
  createdAt?: string;
  datePerformed?: string;
  description?: string | null;
  id?: string;
  mileageAtService?: number;
  notes?: string | null;
  serviceProvider?: string | null;
  updatedAt?: string;
  vehicleId?: string;
}